requirements.txt
requests>=2.31.0
PyYAML>=6.0
pika>=1.1.0
pytest>=6.2.4
coverage>=5.5
redis>=4.5.5
pyzmq>=27.0.0

# SocketIO + deps
python-engineio[client]>=4.11.2
python-socketio[client]>=5.12.1
simple-websocket>=1.1.0
#
bidict>=0.23.1
wsproto>=1.2.0
h11>=0.14.0

# TaskNode
setproctitle>=1.3.3
